requirements.txt
# Backend dependencies for RAG chatbot - Compatible with Python 3.13
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
openai>=1.0.0
pinecone>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.5

# For processing different file types
PyPDF2>=3.0.0
python-docx>=0.8.0
beautifulsoup4>=4.12.0

# For text processing and embeddings
tiktoken>=0.5.0
numpy>=1.24.0